+++ Source/DotSpatial.Data.Rasters.GdalExtension/GdalRaster.cs
@@ -582,6 +582,9 @@ namespace DotSpatial.Data.Rasters.GdalExtension
             }
             double[] affine = new double[6];
             _dataset.GetGeoTransform(affine);
+            // in gdal (row,col) coordinates are defined relative to the top-left corner of the top-left cell
+            // shift them by half a cell to give coordinates relative to the center of the top-left cell
+            affine = (new AffineTransform(affine)).TransfromToCorner(0.5, 0.5);
             ProjectionString = projString;
             Bounds = new RasterBounds(base.NumRows, base.NumColumns, affine);
             PixelSpace = Marshal.SizeOf(typeof(T));